backend/requirements.txt
flask
flask-cors
python-frontmatter
